added support for more than 255 sectors in an ATA read, added logging to ATA

+++ src/kernel/drivers/ata/ata.h
@@ -3,7 +3,7 @@
 
 #define kATA_WaitTimeout 100000
 
-bool ataReadSectors(u32 lba, u8 count, void* dest);
+bool ataReadSectors(u32 lba, u32 count, void* dest);
 
 #ifdef kATA_SelfTest
 void ataSelfTest();
